Nuget packages update (EF Core 2.1) + Lazy loading database issue in Shard (#914)
* Fixed an issue identified in EF 2.1 where context was being modified...

One of the nested loops was lazy loaded, then we iterated through those results pulling more records from the db. This raised an exception in EF 2.1.

The fix is to eager load the results that we iterate over.

Because of this, I've switched everything over to eager loading anyway. We don't do any lazy loading in ACE (we have no requirement to).

* NuGet packages updated + Changelog

* Couple cosmetic changes and null check fixes