CommandHandlerHelper (Yum!) Added (#1027)
There were a ton of commands that worked both on the console, and from a player session.

These commands had conditions like the following to determine if the output should be sent to the console or to the sesion:
if (session != null)
 sendToSession()
else
 SendToConsole()

CommandHandlerHelper cleans up this.

In addition, there were places in the code that were doing Console.WriteLine that should have been using the logger instead.