Action queue remove dequeue (#1094)
* Landblock LoadMeshes summary improved

* Landblock CellLandblock/LandblockInfo get only (set in ctor)

* Fix LScape get_landblock thread safety

* Fix EnchantmentManager Surpass (retval)

* Only create the WorldObject.ObjectGuid once

We don't need to instantinate a new ObjetGuid per get of the property.

* castclass (cast) is faster than isinst (as), also fix double enumeration

* Streamline Attributes and Vitals

They are fixed and set in SetEphemeralValues, called by the ctor.

They do not need the additional layer that Skills do.

* BiotaPropertiesEnchatmentRegistry direct access fixes (Use extensions instead)

* Another BiotaPropertiesEnchantmentRegistry -> Extension fix

* Initial EnchantmentManagerWithCaching

* Added the remaining functions to the cache

* DelayManager using the wrong Min. woops...

Min() was the linq version that enumerates the whole collection. Min is the built in fast version for the SortedSet

* Fix PortalDatDatabase properties

They should only be set once, not executing the getter on every request.

Furthermore, SpellBase was referencing the static DatManager via the Unpack command to construct SpellWords. A dat file type should not be referencing the static datmanager. Instead, we use the GetSPellWords that takes the reference to DatManager to generate the same result.

* Cleanup loadalllandblocks status output

* Improve LoadAllLandblocks comments

* Remove ability to Dequeue from ActionQueues

This is a small change, yet a big change.

It removes a big feature of the ActionQueue system, the ability to dequeue pending acitons. This was added during the time when we had the idea of using micro transactions in ACE via these ActionQueues.

We have since diverged from micro transactions, and thus, the ability to Dequeue is no longer needed. We also have no use case for the ability to dequeue a pending action before it's acted on in the future.

By removing this requirement, we can switch the ActionQueue backing from a LinkedList to a ConcurrentQueue.

This change reduces the CPU ActionQueues require by half.

This adds up when you consider that every WorldObject is an Actor with their own action queues, and those action queues are ticked every UpdateGameWorld tick.