Adding scrolls to loot profiles (#1102)
* Adding scrolls to loot profiles

Scroll weenies are cached into lists, then created from those weenies, so the database isnt hit everytime there is a new scroll created.

* Fixed so that the lists for spell weenies are persistent and only are added when weenies are not present.

This should be good to go as I tested it thoroughly.

* Loads all weenies on first weenie looup

also, renamed variable of  result to spellWeenie and returned the dict value and not the value of the lookup.

* changed field to static so no need to make Database object

* WorldDatabase GetRandomWeeniesOfType -> GetRandomCachedWeeniesOfType

WorldDatabasePrecaching set to true is a requirement for proper loot generation

* Improve getting scroll weenies by spell id

* GetRandomCachedWeeniesOfType -> GetRandomWeeniesOfType