More profiling tweaks (#1242)
* move the IsGenerator check to the base WorldObject HeartBeat

* Move EmoteManager.HeartBeat from WorldObject to Creature

* Change WO HeartBeatInterval check to a NextHeartBeatTime

Suggestion from gmriggs

* check if HasEnchantments before EnchantmentManager.HeartBeat()

This still encounters a lock but still, the overall performance is very good here.

For extreme performance we could expose a non-thread safe HasEnchantmentsForTick function that checks the colleciton size without using a lock.

* switch Monster_Tick to use currentUnixTime instead of DateTime.UtcNow

* Only instantiate actionQueue when it's actually used

This reduces the overhead of calling actionQueue.RunActions() that queries a ConcurrentQueue for it's count.

* Expand /serverstatus report

* Fix ObjCell.FindObjCollisions InvalidOperationException

* use faster Queue.IsEmpty check in ActionQueue.RunActions

* Only set RegenerationInterval if > 0 (per gmriggs)

* Monster_Tick changes (per gmriggs)

* fix QueueFirstHeartbeat

* Add some UpdateGameWorld RateMonitors to /serverstatus

* Remove nullable from NextHeartBeatTime (was not needed)

Comparing a nullable to a non-nullable takes almost 2x as long as a non-nullable to a non-nullable.

* AgressiveInlining suggestions for the compiler