Reuse available Dynamic GUIDs (#1493)
* Reuse available Dynamic GUIDs

On ACE startup, the shard database is queried for available id's in the dynamic guid range.

This helps further reduce (max + 1) id allocation.

It's a bit premature, but, based on current architecture, I don't know if we'll ever need to do a id defragment.

A requirement of this method is that ace_world may only contain guids in the static range.

* GetMaxGuidFoundInRange has no valid use in ace_world

All ids must be in the static range

* Add world update script