Move WorldObject specific managers to a different folder (#2305)
This helps better visualize the manager usage between
- server wide
- world object localized

I'm doing this because as I travel down the rabbit hole of server-wide thread safety, these types of separations help make it easier to understand what managers need more attention for thread safety vs world object localized managers that sould be thread safe because their parent wo should be thread safe (key word *should*)

I'll be adding thread safety to these server-wide managers and want to organize them better first to reduce the size of the multi-threading PR. This makes it easier to read, and easier to roll back if needed.

QuestManager remains in the server wide folder because it is used for both WorldObjects and Fellowships.