ShardDatabaseWithoutCaching (#2569)
This makes the legacy code for bypassing EF tracking for non player biotas back into master, and optional.

This will make it easier for ACE devs to update this legacy code for those servers that still require it, reducing the amount of work they'll have to do to merge to latest when certain schema changes happen.