Support application startup from outside directory (#2588)
In current master, on a windows platform, ACE requires the current environment directory to be the same location that the ACE.Server.exe (.NET core 3.1) exists.

If you try to startup the application from an outside directory, the Config.js and log4net.config will not be found as only the Environment.CurrentDirectory were searched.

Now, that directory is searched first. If the files are not found, the ExecutingAssembly Location is searched next.

This needs to be tested on Linux.