Improve World DB Startup Caching (#3587)
This is significantly faster than what is currently in master, and just slightly slower than PR #3582

It is now a single threaded operation, just like #3582 but instead of stitching the objects together manually, it lets EF stitch the objects together.

EF Tracking is required for the stitching to work.

This is the recommended pattern for these types of actions when dealing with EF.

I think similar improvements can be made to startup player caching