Fix StackOverflow on consecutive syncrhonous OnDataReceive calls (#3958)
There is no guarantee that OnDataReceive will be called asynchronously.

Every time it is called asynchronously, the stack is reset as it's called via a new task.

If it's called synchronously enough times consecutively, the stack will overflow. Ensuring that it's called on a new task in these scenarios will prevent stack overflows.