Cast_On_Strike - Multi-strike, Cleave, and Split arrow
Added bools & properties to allow cast on strike to target cleaved targets, as well as a bool/float to allow multi-strike proc chance on every attack in the sequence instead of only the first successful hit.

This allows content creators flexibility in item creation and the lower the number, the higher the effect for the float. 0.0f = proc chance remains the same. 1.0f = is essentially turning off the proc chance.

SelfTargeted Procs will not trigger, nor will ring spells.
Single target procs only.

Tested with a Healing Tunic and a Tuskers fate - result was as expected.
I do not have any split arrow weenies, did not test functionality here.