Perf: stop ticking dead monsters.
The `Monster_Tick` function exits early on `IsDead`.  There's no reason for us to continue keeping this monster in the sorted creatures list - this just ends up causing additional churn through more iterations, more remove-readd linked list allocations, and more next-tick-time calculations.

There is logic in `Monster_Tick` before the early exit, but none of it is meaningful or shouldn't be skipped:
- `gamePiece.Tick` and `pet.Tick`, which are for creatures that have non-zero health and wouldn't be considered dead
- The calculation for `NextMonsterTickTime`, which is unnecessary if we are not ticking the monster again.
- The check for `!IsAwake` which conditionally takes non-awake monsters and modifies their state and/or wakes or alerts faction mobs, unnecessary for a dead monster.

When we later find the creature in our `pendingRemovals`, we'll call `sortedCreaturesByNextTick.Remove(creature)` - this is a safe operation on a linked list, and will simply not find the creature.