Reduce empty corpse decay time to 3s by default.
This PR replaces the existing static corpse decay time with a property.  It also reduces the default value from 15s to 3s.

This accelerated decay time will impact corpses that are spawned but empty, corpses that are looted to an empty state, and accelerated decay for players who spam corpses in a landblock.