Remove extra item save in rent payment. (#251)
The old code would call TryRemoveFromInventoryWithNetworking against the player (which triggers a save of the item setting the owner to null) followed by a deep save of the slumlord's inventory (which triggers a save of all the items setting the owner to the slumlord).

Our existing code for blocking "races" involves preventing enqueueing a new save when one is already in progress.  In practice, this leads to losing a save, and the save lost was the "set owner to slumlord".  In certain cases, no subsequent save would occur and the transfer of items would be lost.

This change fixes this by deferring the first (removal)'s save, which should make the transfer of ownership atomic - if the transfer fails, the player will retain the item.