Separate Inbound and Outbound Message types. (#276)
This PR makes it much more clear which types of messages may be received and which may be sent.  It also makes it more clear at callsites that an outbound message is being enqueued to be sent to a player.

Future work may leverage this to avoid dictionary lookups on every inbound handler (it's simple enough to use a switch statement instead).