Modify the new life aug curve to use a different formula. (#282)
As discussed in chat, we are moving to using the formula (1.0 - (1.0 - r)^a).  Unlike the existing formulas, this diminishing returns formula always allows equal relative steps at each point.  For example, using the default r value of 0.0034597 means that we always cover half the remaining distance in 200 steps.

The update also stores the old and new values in spell properties, and allows the `new_life_aug_curve_pct` to be ramped up from 0 to 1 to gradually enable the feature.  This will allow us to "turn up the heat" gradually so as to minimize sharp changes.  Further, since the values are interpolated at damage time, the feature can be enabled or disabled without requiring players to rebuff.