Introduce ServerConfig to replace PropertyManager's storage of config properties.
The existing setup is messy - it's key-value lookups in a dictionary, which means every time we check a config value we incur the cost of a lookup.  Replacing this with direct property accesses makes the access more direct and easy to follow and improves efficiency.

Additionally, this removes the use of multiple defaults for a single property in many places (both at the callsite and in the definition) and removes lookups for properties that don't exist.