Calculate StatModValueNew at the time of use. (#289)
This fixes a bug with the hot-swapping between the old and new life curves.

In the current setup, PropertiesEnchantmentRegistry.StatModValueNew is only calculated at the time of cast.  Since this is persisted to the database, and there is no field for StatModValueNew, a server restart will attempt to restore these enchantments and will initialize `StatModValueNew` as 0.  This has the effect of the new curve mod never being populated and thus 0 being used as a damage multiplier, making the player immune.

We fix this by using the `AugmentationLevelWhenCast` field to calculate the `StatModValueNew` at the time of first use.  This means that the value will always be calculated correctly from its first use.

Since the values are cached in the EnchantmentManagerWithCaching, that means we only do this calculation once, at the time of first use. Thus, it is not expected to be a performance issue.