Add EnumFlagRandom for randomly selecting from an Enum flag. (#295)
This PR introduces the EnumFlagRandom utility to provide an efficient LINQ-free way to randomly select a single valid value from an enum flag.  This new method uses bitwise operations and a static cache to determine valid enum flags at runtime.

We then use this new functionality in two frequently-used code paths:
- Monster_Awareness.cs, where we use it to efficiently select a targeting method from the available set.
- DamageType.cs, where using it allows us to inline the damage selection logic into Player_Combat.cs and Creature_Combat.cs, removing the "magic" recursion of the old SelectDamageType method.

This change improves performance by reducing memory allocations (no heavyweight LINQ/ToList as in EnumHelper.GetFlags) and eliminates the possibility of an index-out-of-bounds when accessing an element in the list by index.